Return relay handle to JS
The handle contains a pointer back to the reference-counted relay selector instance.  This lets us implement the `Drop` trait so JS users don't have to explicitly free memory.  When the JS garbage collector frees the handle, we expect the `drop()` method to be invoked in Rust, which decrements the reference count to the selector.  When all references to the selector have been freed, we can perform cleanup actions on the selector itself.